Fix: move method and correct functionality
- As suggested by @divmain the method was moved into a separate
  mixin for reuse, as well as, splitted into smaller parts

- Use `git branch --merged` instead of `--no-merged. This fixes a
  bug where the most obvious branches are not returned since no-merged
  means that only branches that are "unreachable" from given node
  are returned, and we actually want the opposite.
